1.补充代码建议

+++ m-parent/m-base/src/main/java/com/mmmail/base/module/support/heartbeat/core/HeartBeatRunnable.java
@@ -53,7 +53,7 @@ public class HeartBeatRunnable implements Runnable {
         RuntimeMXBean runtimeMXBean = ManagementFactory.getRuntimeMXBean();
        this.projectPath = System.getProperty("user.dir");
        this.serverIps = new ArrayList<>(NetUtil.localIpv4s());
-       this.processNo = Integer.valueOf(runtimeMXBean.getName().split("@")[0]).intValue();
+       this.processNo = Integer.valueOf(runtimeMXBean.getName().split("@")[0]);
        this.processStartTime = LocalDateTime.ofInstant(Instant.ofEpochMilli(runtimeMXBean.getStartTime()), ZoneId.systemDefault());
     }
